improve UI latency feedback.
