+++ views/list.pug
@@ -171,6 +171,10 @@ block append head
     if !min
         link(rel='stylesheet', href= '/users/list/css')
 
+block append scripts
+    script(type="module")
+        include ../src/js/edit/feedback.js
+
 block content
     include util.pug
     .tabletop
